refactor(sdk): encapsulate nonce cache into dedicated NonceCache struct
Address PR #3111 review comments from shumkov:
- Replace NonceCacheEntry tuple with named struct (current_nonce,
  last_fetch_timestamp, last_fetched_platform_nonce)
- Replace (Identifier, Identifier) key with IdentityContractPair struct
- Create NonceCache struct encapsulating all nonce logic with per-map
  Mutex locking and configurable stale time
- Remove InternalSdkCache (was nonce-only, pointless wrapper)
- Make Sdk nonce methods thin wrappers delegating to NonceCache
- Move get_or_fetch_nonce, get_current_time_seconds,
  DEFAULT_IDENTITY_NONCE_STALE_TIME_S, and tests to internal_cache/mod.rs
- Remove LastQueryTimestamp and StalenessInSeconds type aliases from sdk.rs

Public API unchanged — all callers continue using sdk.get_identity_nonce(),
sdk.get_identity_contract_nonce(), and sdk.refresh_identity_nonce().
